requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
emoji==1.2.0
idna==2.10
PySocks==1.7.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
emoji==1.2.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.2.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
